Move MR generation code to separate file
